reverting some changes since rpc_test.go is not running in CI
