+++ gateway/rpc_test.go
@@ -4,12 +4,10 @@
 package gateway
 
 import (
-	"errors"
 	"net/http"
+	_ "net/http"
 	"testing"
 	"time"
-
-	"github.com/stretchr/testify/assert"
 
 	"github.com/TykTechnologies/gorpc"
 	"github.com/TykTechnologies/tyk/apidef"
@@ -17,6 +15,48 @@ import (
 	"github.com/TykTechnologies/tyk/rpc"
 	"github.com/TykTechnologies/tyk/test"
 )
+
+func StartSlaveGw(connectionString string, groupId string) *Test {
+	conf := func(globalConf *config.Config) {
+		globalConf.SlaveOptions.UseRPC = true
+		globalConf.SlaveOptions.RPCKey = "test_org"
+		globalConf.SlaveOptions.APIKey = "test"
+		globalConf.Policies.PolicySource = "rpc"
+		globalConf.SlaveOptions.GroupID = groupId
+		globalConf.SlaveOptions.CallTimeout = 1
+		globalConf.SlaveOptions.RPCPoolSize = 2
+		globalConf.AuthOverride.ForceAuthProvider = true
+		globalConf.AuthOverride.AuthProvider.StorageEngine = "rpc"
+		globalConf.SlaveOptions.ConnectionString = connectionString
+	}
+	return StartTest(conf)
+}
+
+func startRPCMock(dispatcher *gorpc.Dispatcher) (*gorpc.Server, string) {
+
+	rpc.GlobalRPCCallTimeout = 100 * time.Millisecond
+
+	server := gorpc.NewTCPServer("127.0.0.1:0", dispatcher.NewHandlerFunc())
+	list := &customListener{}
+	server.Listener = list
+	server.LogError = gorpc.NilErrorLogger
+
+	if err := server.Start(); err != nil {
+		panic(err)
+	}
+
+	return server, list.L.Addr().String()
+}
+
+func stopRPCMock(server *gorpc.Server) {
+
+	if server != nil {
+		server.Listener.Close()
+		server.Stop()
+	}
+
+	rpc.Reset()
+}
 
 const apiDefListTest = `[{
 	"api_id": "1",
@@ -105,10 +145,10 @@ func TestSyncAPISpecsRPCFailure_CheckGlobals(t *testing.T) {
 		return `[]`, nil
 	})
 
-	rpcMock, connectionString := startRPCMock(t, dispatcher)
-	defer stopRPCMock(t, rpcMock)
+	rpcMock, connectionString := startRPCMock(dispatcher)
+	defer stopRPCMock(rpcMock)
 
-	ts := StartSlaveGw(t, connectionString, "")
+	ts := StartSlaveGw(connectionString, "")
 	defer ts.Close()
 
 	store := RPCStorageHandler{Gw: ts.Gw}
@@ -147,10 +187,10 @@ func TestSyncAPISpecsRPCSuccess(t *testing.T) {
 	})
 
 	t.Run("RPC is live", func(t *testing.T) {
-		rpcMock, connectionString := startRPCMock(t, dispatcher)
-		defer stopRPCMock(t, rpcMock)
+		rpcMock, connectionString := startRPCMock(dispatcher)
+		defer stopRPCMock(rpcMock)
 
-		ts := StartSlaveGw(t, connectionString, "")
+		ts := StartSlaveGw(connectionString, "")
 		defer ts.Close()
 
 		GetKeyCounter = 0
@@ -213,7 +253,7 @@ func TestSyncAPISpecsRPCSuccess(t *testing.T) {
 		if GetKeyCounter != 0 {
 			t.Error("getKey should have been called 0 times")
 		}
-		stopRPCMock(t, nil)
+		stopRPCMock(nil)
 	})
 
 	t.Run("RPC is back, hard reload", func(t *testing.T) {
@@ -236,10 +276,10 @@ func TestSyncAPISpecsRPCSuccess(t *testing.T) {
 			return jsonMarshalString(CreateStandardSession()), nil
 		})
 		// Back to live
-		rpcMock, connectionString := startRPCMock(t, dispatcher)
-		defer stopRPCMock(t, rpcMock)
+		rpcMock, connectionString := startRPCMock(dispatcher)
+		defer stopRPCMock(rpcMock)
 
-		ts := StartSlaveGw(t, connectionString, "")
+		ts := StartSlaveGw(connectionString, "")
 		defer ts.Close()
 
 		time.Sleep(1000 * time.Millisecond)
@@ -263,9 +303,9 @@ func TestSyncAPISpecsRPCSuccess(t *testing.T) {
 	})
 
 	t.Run("RPC is back, live reload", func(t *testing.T) {
-		rpcMock, connectionString := startRPCMock(t, dispatcher)
+		rpcMock, connectionString := startRPCMock(dispatcher)
 
-		ts := StartSlaveGw(t, connectionString, "")
+		ts := StartSlaveGw(connectionString, "")
 		defer ts.Close()
 		time.Sleep(100 * time.Millisecond)
 
@@ -321,10 +361,10 @@ func TestSyncAPISpecsRPC_redis_failure(t *testing.T) {
 		return jsonMarshalString(CreateStandardSession()), nil
 	})
 
-	rpcMock, connectionString := startRPCMock(t, dispatcher)
-	defer stopRPCMock(t, rpcMock)
+	rpcMock, connectionString := startRPCMock(dispatcher)
+	defer stopRPCMock(rpcMock)
 
-	ts := StartSlaveGw(t, connectionString, "")
+	ts := StartSlaveGw(connectionString, "")
 	defer ts.Close()
 
 	t.Run("Should load apis when redis is down", func(t *testing.T) {
@@ -397,41 +437,4 @@ func TestOrgSessionWithRPCDown(t *testing.T) {
 	if found {
 		t.Fatal("org  session should be null:")
 	}
-}
-
-func TestRPCStorageHandler_Disconnect(t *testing.T) {
-	rpc.UseSyncLoginRPC = true
-
-	t.Run("disconnect ok", func(t *testing.T) {
-		dispatcher := gorpc.NewDispatcher()
-		dispatcher.AddFunc("LoginWithGroup", func(clientAddr string, groupData *apidef.GroupLoginRequest) bool {
-			return true
-		})
-		dispatcher.AddFunc("Disconnect", func(clientAddr string, groupData *apidef.GroupLoginRequest) error {
-			return nil
-		})
-
-		rpcMock, connectionString := startRPCMock(t, dispatcher)
-		defer stopRPCMock(t, rpcMock)
-
-		ts := StartSlaveGw(t, connectionString, "test-group")
-		defer ts.Close()
-
-		r := &RPCStorageHandler{Gw: ts.Gw}
-		connected := r.Connect()
-		assert.True(t, connected)
-
-		err := r.Disconnect()
-		assert.Nil(t, err)
-	})
-	t.Run("disconnect error", func(t *testing.T) {
-		ts := StartTest(nil)
-		defer ts.Close()
-
-		r := &RPCStorageHandler{Gw: ts.Gw}
-
-		err := r.Disconnect()
-		expectedErr := errors.New("RPCStorageHandler: rpc is either down or was not configured")
-		assert.EqualError(t, err, expectedErr.Error())
-	})
 }
